+++ changelog.md
@@ -24,6 +24,7 @@
 - Bugfix: The tuple redistribution strategy was not able to handle local destinations 
 - Bugfix: Fixed NPE during bounding box creation
 - Bugfix: Deleted tuples were read form SSTables without the original timestamp
+- Bugfix: Duplicate tuples (caused by queries on replicates) are filtered in the client code
 
 ### Version 0.2.0 (Beta-2) - 28.11.2016
 - New Feature: Java 8 is now required to build the project